requirements.txt
backcall==0.2.0
blinker==1.4
CacheControl==0.12.5
cachetools==3.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
firebase-admin==3.2.0
Flask==1.1.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.7.2
google-auth-httplib2==0.0.3
google-cloud-core==1.1.0
google-cloud-firestore==1.6.0
google-cloud-storage==1.23.0
google-crc32c==0.1.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.31.0
httplib2==0.14.0
idna==2.10
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jupyter-client==6.1.6
jupyter-core==4.6.3
lxml==4.4.2
MarkupSafe==1.1.1
msgpack==0.6.2
oauth2client==4.1.3
parso==0.7.0
passlib==1.7.2
pickleshare==0.7.5
pika==1.1.0
prompt-toolkit==3.0.5
protobuf==3.11.1
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==228
pyzmq==19.0.1
requests==2.24.0
retrying==1.3.3
rsa==4.0
six==1.15.0
spyne==2.12.16
SQLAlchemy==1.3.18
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.0
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
